Update core terminal display logic & features
The script has been updated to extend its capabilities. It now uses
ascii escapes to draw to a buffered window to improve video performance
- smoother and less lag on redraws - while also adding new features.
This include the ability to pass a file to the shell command such as
a mp4 or gif - video input - and have its frame being generated without
having to run the ffmpeg command before. The script also checks for
the input video original frame rate for better display performance.
[todo]
- Need to improve the core wait logic; sleep is ok for now and get the
job done but its inefficient at best and possibily the source of some
quirks with the animation
- Seprate options for jp2a in a proper variable and maybe pass it as
an argument when running the script, or fall back to some default settings
if no argument is present.
:3