Implement new frame wait logic + Debug utils
Alternative wait logic for frame times, old sleep one is just commented
out.
This has been implemented in hopes of fixis a current bug, where the video playback
is delayed in case of videos at 60fps with many frams but speed up in the case
of gifs at 25/30 fps with less frames. Some utils have been created to help with
debugging the issue but little progress has been made.
Ideally having the frame gen in the loop isn't great, should instead pre-process
the frames in a buffer queue and then print the frames at playback separating
rendering and playback but for sake of simplicity this is the way for now.
Maybe bash isn't the best choice for this setup but for a simply proof of concept
this will have to do.
the new rendering system basically waits for a frame time to arrive and only if the
time is correct will print it but still it doesn't seem to be behaving correctly,
maybe its the loop fault or it could simply be the hoverhead of having to perform
all these action and have the video printed by that time.