Add ascii-display.c basic logic
the algorithm does the same base thing of ascii-animation.sh but now in C!
It allows for some better manipulation of the logic even if a bit more
complicated in logic and syntax.
At its core the algo first runs ffmpeg to disect the input video into
its frames as png - see frame-generator.sh - converts each frame to
a ascii escape sequence and puts it into the frames.txt file.
the c part just reads each frame at a time, prints it, waits for a
set frame-time (here its manual for now, will try to implement a self
detecting system) clears the screen and repeats untill the file is no more.
everything is suboptimal for now, will have to improve, but its a good proof
of concept for now :)