Add fps calculation, frame time sleep + refactor
refactor is just concerning the functions - following best practices -
and renaming file_name to video_name, more clarity on context.
fps calculation is done the same way as before with ffprobe and bc,
just using some odd c to send the command and retrieve its content.
for the frame time based sleep, usleep is used, maybe not the best
solution but good for getting thins started. it could be good to also
try and implement the old-new system of wating according to the video
playback instead of a fixed frame time.